Fix special case of certain longish comments.
Usually, 0 0 17 3 nn$ represents the start of a comment broken up into
multiple strings.  But in a few cases (I think when nn is 67), it's still
a single string.  I *think* I've fixed the problem.

But I have new problems, such as some missing text.  We'll see what's up
with that next.